+++ src/index.js
@@ -5,7 +5,9 @@ const queryErrorHandler = require('./handlers/queryErrorHandler.js')
 const depositHandler = require('./handlers/depositHandler.js')
 const depositErrorHandler = require('./handlers/depositErrorHandler.js')
 const balanceHandler = require('../src/handlers/balanceHandler.js')
-const balanceErrorHandler = require('../src/handlers/balanceErrorHandler.js')
+const balanceErrorHandler = require('./handlers/balanceErrorHandler.js')
+const withdrawHandler = require('./handlers/withdrawHandler.js')
+const withdrawErrorHandler = require('./handlers/withdrawErrorHandler.js')
 
 exports.handler = async (event, context) => {
     logEvent(event);
@@ -37,6 +39,14 @@ exports.handler = async (event, context) => {
             body = await balanceHandler.handle(event)
         } catch (err) {
             const errorHandlerResp = await balanceErrorHandler.handle(err)
+            statusCode = errorHandlerResp.statusCode
+            body = errorHandlerResp.body
+        }
+    } else if (event.requestContext.routeKey == 'POST /xp/withdraw') {
+        try {
+            body = await withdrawHandler.handle(event)
+        } catch (err) {
+            const errorHandlerResp = await withdrawErrorHandler.handle(err)
             statusCode = errorHandlerResp.statusCode
             body = errorHandlerResp.body
         }