+++ src/service/rconService.js
@@ -31,6 +31,14 @@ async function removeXpPoints(rconClient, userId, amount) {
     await parseRemoveXpPointsResp(serverResp)
 }
 
+async function addXpPoints(rconClient, userId, amount) {
+    const cmdString = `xp add ${userId} ${amount} points`
+    await rconClient.authenticate(SERVER_RCON_PASS)
+    const serverResp = await rconClient.execute(cmdString)
+    rconClient.disconnect()
+    await parseAddXpPointsResp(serverResp)
+}
+
 async function parseQueryResp(serverResponse) {
     if (serverResponse == rconServiceErrorEnum.NO_PLAYER_FOUND) {
         logger.log('[rconService] ' + serverResponse)
@@ -50,6 +58,14 @@ async function parseRemoveXpPointsResp(serverResponse) {
     }
 }
 
+async function parseAddXpPointsResp(serverResponse) {
+    if (serverResponse == rconServiceErrorEnum.NO_PLAYER_FOUND) {
+        logger.log('[rconService] ' + serverResponse)
+        throw Error(rconServiceErrorEnum.NO_PLAYER_FOUND)
+    }
+}
+
 exports.queryXpLevels = queryXpLevels
 exports.queryXpPoints = queryXpPoints
 exports.removeXpPoints = removeXpPoints
+exports.addXpPoints = addXpPoints