+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project attempts to conform to [SemVer](https://semver.org/) guidelines whe
 ## [v0.2.0]
 * [Implementing Deposit XP API](https://github.com/Ubunfu/mc-xp-bank/pull/1)
 * [Implementing XP Account Balance API](https://github.com/Ubunfu/mc-xp-bank/pull/2)
+* [Implementing Withdraw XP API](https://github.com/Ubunfu/mc-xp-bank/pull/3)
 
 ## [v0.1.0]
 * Implementing Query XP API